requirements.txt
alfacoins-api-python==0.1.2
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-model-utils==3.2.0
idna==2.8
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
django-compressor-parceljs==2.3.1